Bug fixed
Fix comments. added id check and fix problem with idempotence, check created slots for coach

+++ PMFightAcademy.Admin/Controllers/BookingController.cs
@@ -21,13 +21,15 @@ namespace PMFightAcademy.Admin.Controllers
     public class BookingController : ControllerBase
     {
         private readonly IBookingService _bookingService;
+        private readonly IWorkWithIdService _checkId;
 
         /// <summary>
         /// Constructor for booking
         /// </summary>
-        public BookingController(IBookingService bookingService)
+        public BookingController(IBookingService bookingService,IWorkWithIdService checkId)
         {
             _bookingService = bookingService;
+            _checkId = checkId;
         }
 
         ///// <summary>
@@ -105,6 +107,11 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetBookedServiceForClient(int clientId, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(clientId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var bookings = await _bookingService.TakeBookingOnClient(clientId);
             if (bookings.Any())
             {
@@ -132,6 +139,11 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetBookedServiceForCoach(int coachId, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(coachId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var bookings = await _bookingService.TakeBookingForCoach(coachId);
 
             if (bookings.Any())
@@ -151,7 +163,7 @@ namespace PMFightAcademy.Admin.Controllers
         /// <see cref="HttpStatusCode.OK"/>return if book is successful deleted
         /// <see cref="HttpStatusCode.NotFound"/> not founded slots</returns>
         /// <remarks>
-        /// Use for delete book , using post method
+        /// Use for delete book 
         /// return ok if successes
         /// return Not found if not founded 
         /// </remarks>
@@ -160,6 +172,11 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> DeleteBook( int bookingId, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(bookingId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var deleted = await _bookingService.RemoveBooking(bookingId, cancellationToken);
 
             if (deleted)
@@ -167,11 +184,11 @@ namespace PMFightAcademy.Admin.Controllers
                 return Ok();
             }
 
-            return NotFound("No Booking");
+            return NotFound("No Booking with suck id");
         }
 
         /// <summary>
-        /// Update file
+        /// Update booking
         /// </summary>
         /// <param name="newBooking"></param>
         /// <param name="cancellationToken"></param>
@@ -194,7 +211,7 @@ namespace PMFightAcademy.Admin.Controllers
                 return Ok();
             }
 
-            return NotFound("No Booking");
+            return NotFound("No Booking ");
         }
 
 
@@ -260,6 +277,11 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetBookedServiceForClientOnDate(int clientId, string dateStart, string dateEnd, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(clientId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var bookings = await _bookingService.TakeBookingForClientOnDate(clientId,dateStart,dateEnd);
             if (bookings.Any())
             {
@@ -290,6 +312,11 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetBookedServiceForCoachOnDate(int coachId,string dateStart,string dateEnd, CancellationToken cancellationToken)
         {
+            if (!_checkId.IsCorrectId(coachId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var bookings = await _bookingService.TakeBookingForClientOnDate(coachId, dateStart, dateEnd);
 
             if (bookings.Any())
