+++ PMFightAcademy.Admin/Controllers/SlotsController.cs
@@ -20,13 +20,15 @@ namespace PMFightAcademy.Admin.Controllers
     public class SlotsController : ControllerBase
     {
         private readonly ISlotService _slotService;
+        private readonly IWorkWithIdService _checkId;
 
         /// <summary>
         /// Slots controller
         /// </summary>
-        public SlotsController(ISlotService slotService)
+        public SlotsController(ISlotService slotService , IWorkWithIdService checkId)
         {
             _slotService = slotService;
+            _checkId = checkId;
         }
 
         //#region Maded Pagination but not used by JS (TILT)
@@ -164,8 +166,12 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetSlotsForCoach([FromRoute] int coachId)
         {
-            
-               var  slots = await _slotService.TakeSlotsForCoach(coachId);
+            if(!_checkId.IsCorrectId(coachId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
+            var  slots = await _slotService.TakeSlotsForCoach(coachId);
 
                if (slots.Any())
                {
@@ -251,6 +257,12 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> DeleteSlots( int slotId,CancellationToken cancellationToken)
         {
+
+            if (!_checkId.IsCorrectId(slotId))
+            {
+                return BadRequest("incorrect Id");
+            }
+
             var deleted = await _slotService.RemoveSlot(slotId, cancellationToken);
 
             if (deleted)
@@ -277,6 +289,10 @@ namespace PMFightAcademy.Admin.Controllers
         [ProducesResponseType(typeof(string), (int)HttpStatusCode.NotFound)]
         public async Task<IActionResult> GetSlotsForCoachFromDateToDate(int coachId,string dateStart,string dateEnd)
         {
+            if (!_checkId.IsCorrectId(coachId))
+            {
+                return BadRequest("incorrect Id");
+            }
 
             var slots = await _slotService.TakeSlotsForCoachOnDates(coachId,dateStart,dateEnd);
